move ts files to src and test
